Import cost-based prediction from rav1e
For now, this is a copy of the code that exists in rav1e.
This is sub-optimal, but no interface exists in rav1e
to allow us to use the functions needed from there.

Since this is currently functional,
merging this as-is and coming back to refactor later
seems to be the approach.