Add support for VapourSynthDecoder custom environment and arguments (#182)
The `vapoursynth-rs` crate supports 3 inputs: a VapourSynth script file
path, a VapourSynth script as a string, and an `Environment` which can
be evaluated beforehand to contain an output. `VapourSynthDecoder` only
supported the first option and also does not set script variables
(VSPipe arguments).

This PR adds support for the other 2 inputs and accepts a
`HashMap<String, String>` of script variables to set before evaluating a
VapourSynth script. There is a single function that sets the variables
for any given `Environment` and it is used for file and script inputs.
Note: The `new()` function signature has changed from `source: &Path` to
`env: Environment` allowing users to supply a custom `Environment`. The
previous function has been renamed to `new_from_file(source: &Path,
arguments: Option<HashMap>)`. The function signature changed due to
adding `arguments` so a rename for more descriptive functions made
sense.

This functionality is necessary for feature request rust-av/Av1an#994
and to fix Issue rust-av/Av1an#1029.

Thank you,
\- Boats M.

---------

Co-authored-by: Boats McGee <{ID}+{username}@users.noreply.github.com>
Co-authored-by: Boats McGee <BoatsMcGee@users.noreply.github.com>
Co-authored-by: Josh Holmer <jholmer.in@gmail.com>